doccer_requirements.txt
click==8.0.1
docstring-parser==0.10
ghp-import==2.0.2
importlib-metadata==4.8.1
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs-material-extensions==1.0.3
packaging==21.0
Pygments==2.10.0
pymdown-extensions==8.2
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
pyyaml_env_tag==0.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
watchdog==2.1.5
zipp==3.5.0
